Rename HttpBase to BasicHttp to avoid confusion
