+++ src/autov-cod/http_server.ts
@@ -12,23 +12,23 @@ import {
 import { CondPromise } from '../util/cond_promise.js'
 import {
     CONTENT_TYPE_JSON,
-    HttpBase,
+    BasicHttp,
     HttpStatus,
     normalizeUrl,
-} from '../common-cod/http_base.js'
+} from '../common-cod/basic_http.js'
 import { TrialModel } from '../common-cod/trial_model.js'
 
 export class HttpServer {
     trialModel: TrialModel
     config: Config
-    base: HttpBase
+    base: BasicHttp
     // A promise that resolves after we receive a 'close' event from http.Server
     serverFinished: CondPromise
 
     constructor(trialModel: TrialModel, config: Config) {
         this.trialModel = trialModel
         this.config = config
-        this.base = new HttpBase(config.toHttpConfig())
+        this.base = new BasicHttp(config.toHttpConfig())
         this.serverFinished = new CondPromise()
     }
